+++ actionview/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Rails 6.0.0.alpha (Unreleased) ##
 
+*   Fix finding dependencies for templates. No longer find dependencies inside
+    of HTML tags and also calls to render that include a instantiating an object.
+
+    Fixes #21951 and #23536
+
+    *Daniel Fox*
+
 *   Change translation key of `submit_tag` from `module_name_class_name` to `module_name/class_name`.
 
     *Rui Onodera*